Start pulling out utils that might be causing a circular dependency error
