+++ playground/src/utils/index.ts
@@ -11,9 +11,11 @@ import { twMerge } from "tailwind-merge";
 export * from "./screen-size";
 export * from "./otel-helpers";
 export * from "./vendorify-traces";
+export { isSensitiveEnvVar } from "./env-vars";
 export { renderFullLogMessage } from "./render-log-message";
 export { truncateWithEllipsis } from "./truncate";
 export { parseEmbeddedConfig } from "./config-parser";
+export { safeParseJson } from "./safe-parse-json";
 export function formatDate(d: Date | string) {
   return format(new Date(d), "HH:mm:ss.SSS");
 }
@@ -196,16 +198,6 @@ export function formatHeaders(headers: Record<string, string>): string {
     .join("\n");
 }
 
-export const safeParseJson = (jsonString: string) => {
-  try {
-    const parsed = JSON.parse(jsonString);
-    return parsed;
-  } catch (error) {
-    console.error("Failed to parse JSON:", error);
-    return null;
-  }
-};
-
 export function getHttpMethodTextColor(method: string) {
   return {
     GET: "text-info",
@@ -217,28 +209,6 @@ export function getHttpMethodTextColor(method: string) {
     HEAD: "text-info",
     WS: "text-success",
   }[String(method).toUpperCase()];
-}
-
-export function isSensitiveEnvVar(key: string) {
-  if (!key) {
-    return false;
-  }
-
-  return (
-    key.includes("APIKEY") ||
-    key.includes("API_KEY") ||
-    key.includes("ACCESS") ||
-    key.includes("AUTH_") ||
-    key.includes("CREDENTIALS") ||
-    key.includes("CERTIFICATE") ||
-    key.includes("PASSPHRASE") ||
-    key.includes("DATABASE_URL") ||
-    key.includes("CONNECTION_STRING") ||
-    key.includes("SECRET") ||
-    key.includes("PASSWORD") ||
-    key.includes("PRIVATE") ||
-    key.includes("TOKEN")
-  );
 }
 
 export function constructPlaygroundBody(bodyValue: string): PlaygroundBody {
