+++ playground/src/utils/safe-parse-json.ts
@@ -1,0 +1,9 @@
+export const safeParseJson = (jsonString: string) => {
+  try {
+    const parsed = JSON.parse(jsonString);
+    return parsed;
+  } catch (error) {
+    console.error("Failed to parse JSON:", error);
+    return null;
+  }
+};
